==> s5_2_evaluation_manual.py <==
domain_pair_in_context = '''Given the following information as context:

User finickyWasp8: It seems that luscepited-000 is true.
User scornfulPear1: It's possible that luscepited-000 is true.
User sugarySwift8: Arguably, luscepited-000 is true.
User ardentPaella8: I think luscepited-000 is true.
User unhappyBuck7: There is a chance that luscepited-000 is true.
User crushedCockatoo4: It is possibly true that luscepited-000 is true.
User sugaryRhino0: I'm unsure if luscepited-000 is true.
User jealousEggs3: I'm unsure if luscepited-000 is true.
User sheepishBass8: I suspect luscepited-000 is true.
User cheerfulOryx6: It seems that luscepited-000 is true.
User shamefulLeopard1: It's possible that luscepited-000 is true.
User wornoutLeopard4: luscepited-000 might be true.
User culturedShads2: It's possible that luscepited-000 is true.
User wingedCrackers8: It is possibly true that luscepited-000 is true.
User adoringViper8: I think luscepited-000 is true.
User cockyBustard3: luscepited-000 might be true.
User jealousSeafowl0: luscepited-000 might be true.
User artisticSwift4: I'm unsure if luscepited-000 is true.
User crummyPorpoise3: It is possibly true that luscepited-000 is true.
User forsakenGnu2: It's possible that luscepited-000 is true.
User grizzledGnu6: I'm unsure if luscepited-000 is true.
User mellowPie6: luscepited-000 might be true.
User alertMackerel6: It seems that luscepited-000 is true.
User pleasedDoves0: I'd say luscepited-000 is true.
User needfulHoopoe4: It could be that luscepited-000 is true.
User sincereLapwing5: luscepited-000 might be true.
User annoyedCaribou3: It is possibly true that luscepited-000 is true.
User curiousSalami0: I think luscepited-000 is true.
User panickySeahorse2: I think luscepited-000 is true.
User fondCurlew7: It is possibly true that luscepited-000 is true.
User jubilantLlama4: Perhaps luscepited-000 is true.
User worldlySheep2: I'm unsure if luscepited-000 is true.
User tautTeal6: Assuredly, luscepited-000 is true.
User trustingOtter9: I think luscepited-000 is true.
User outlyingLlama0: luscepited-000 might be true.
User thrilledMare3: It's possible that luscepited-000 is true.
User murkyPiglet0: It is possibly true that luscepited-000 is true.
User aboardEland3: I'm unsure if luscepited-000 is true.
User sheepishFlamingo3: It seems that luscepited-000 is true.
User sheepishChile6: I suspect luscepited-000 is true.
User abjectRice0: It seems that luscepited-000 is true.
User lazyCheetah0: It could be that luscepited-000 is true.
User enragedCoconut0: It could be that luscepited-000 is true.
User artisticCaribou4: It's possible that luscepited-000 is true.
User cockyViper5: I'm unsure if luscepited-000 is true.
User hushedPoultry2: I believe luscepited-000 is true.
User jumpyRuffs1: It's possible that luscepited-000 is true.
User betrayedDove8: Assuredly, luscepited-000 is true.
User scornfulPorpoise6: I think luscepited-000 is true.
User exactingRuffs5: luscepited-000 might be true.

How confident are users about luscepited-000 in general?'''
